Support both `/ping-200` and `/ping-200/{any child path}` for simple cases
The existing `/ping-200` endpoint is good for anything that needs a
simple 200 response, but often we want to configure more than one
in a project and it can get confusing which of the services pointing
at `/ping-200` were called.

So add matching for *anything* under `/ping-200` to allow us to namespace
requests without having to add custom handlers at all.